BitmapFont: fix nbsp, ellipsis, support extended ASCII (code page 437)
This fixes the weird character showing up between damage meters and item
names with item health bars enabled.

This also adds support for Unicode ellipsis to be displayed properly, as
well as a bunch of other symbols, by converting incoming Unicode to the
equivalent in the BitmapFont.  This could be used by e.g. mods that want
to change terrain characters to ones that exist in a BitmapFont.

Supported extra characters are the ones in extended ASCII code page 437.
For more info, see: https://en.wikipedia.org/wiki/Code_page_437