Tell position/awareness of creature in interruption prompt
When a new creature appears in the player's field of view, provide its
position and whether it's aware of the player, e.g.

"tough zombie spotted (25N 10E, aware of you)! Stop aiming?"

This helps a lot, since the player can't scroll the view around while
the prompt is up.