Remove fresh water from forests and swamps
This should give more incentive to seek out real water sources, or use
alternatives like toilets in houses and using funnels to catch rain.

I didn't really want to remove fresh water from swamps per se, since it
was present in older versions of the game, but in those older versions,
swamps had swarms of angry mutant bugs to guard the water, which are
barely present anymore, so removal it is.