BitmapFont: Support white-on-black font images with grayness for alpha
This commit adds support for a new kind of bitmap font image:
white-on-black where pixels that are neither fully black or white
represent per-pixel alpha.  The old non-blended white-on-magenta format
is still supported; the BitmapFont loading code checks the top-left
pixel for magenta to decide how to load in the font image.

I made this because I didn't like how the game rendered my preferred
font all spidery and distorted.  With this change however, I can take a
screenshot of my terminal emulator rendering my preferred font the way I
like it, edit it and use it in the game for real.