Report "Sound" correctly when making noise in a moving vehicle
Now when the player fires a gun out of a moving vehicle, the noise made
is reported correctly.

The "Sound" value in the sidebar is based on noises at or next to the
player's position, but they're often made before vehicles, monsters,
NPCs, etc. move, and when a vehicle moves, it moves the player away from
the sound with it, even if the player was the one who made the sound.
This commit fixes this by checking sounds at the moment they're made to
see if they should included in the "Sound" reporting.