Audit gun volumes and weights
This applies all of gun volume and weight corrections as of upstream
commit 5c118cddba6c2d8effbc2420851b70bc1c6a7587 .

Some `magazine_well` values were adjusted too, though the upstream
values had to be rounded up to legacy pre-nested container values.

The Uzi hasn't been audited yet, so its volume has been roughly adjusted
to be similar to the MP5 for the time being.