Reset cursor when changing filter in item picker
This notably fixes being left on a blank page if you were anywhere but
on the first page when in e.g. the eat menu and applied a filter that
had only one page worth of items to choose from.

Figuring out how to do this took way too long, since there's no single
`page` variable to work with; rather, each inventory column has its own
idea of what page it's supposed to be on, and... I don't know the rest,
I just stumbled across code to handle pressing Home and cribbed off of
that.