Always show reload menu for guns that take magazines and empty containers
In these two cases, it was easy to reload and not know what they were
reloaded with, since the menu was typically skipped if only one option
was available.  This should give prior warning before reloading guns
with a magazine that has less ammo than what's currently loaded, for
example.