Reduce default ammo loudness by 40%
Both the code for gun firing sound effects and the hand-entered loudness
values for ammo suggest that the threshold for the loudest firing noise
was supposed to max out at just below 200.  But in testing, I found that
.308 was making 280+ noise.  Reducing the default ammo loudness by 40%
brings this down to about 170, where it makes a "blam" noise instead of
"kerblam".

This also makes suppressors more significant for lower caliber
ammunition types, to offset the downsides of suppressor use in general,
like obscuring sights.