Adjust "bang!" vs. "blam!" gun noise threshold from 150 to 120
Considering the reality bubble size is about 60 spaces in all directions
around the player, a noise of 120 loudness will generate enough interest
from creatures at the edge to walk all the way to the source with no
other stimulus, meaning that a player that makes such a noise and stays
in place will have to fight basically everything in the current reality
bubble.  150, on the other hand, was pretty arbitrary.