Remove nail/flechette loudness, make nail gun/rifle a bit louder
This brings back the old "plink!" and "Brrrip!" messages that they used
to have in older versions of the game.  Otherwise, flechettes were
making noise louder than 9x19mm being fired out of a Glock!

Also make the nail gun and rifle a touch louder.  9 loudness is about
the volume of closing a door, so we'll go with that.